Removed the categories on UIImage, UIImageView as well as the class using them AFImageRequestOperation.
The rationale behind this is simple, BugSense is a crash reporting tool using the AFNetworking library without the apparent need to include categories on UIImage, UIImageView and the class AFImageRequestOperation. Further more, this was the cause of a conflict with AFNetworking included as a dependency in our project, which is likely a different version of the library.